Correccion de atributos de la clase album y correccion del diagrama de clases

+++ Desafio2-Code/album.cpp
@@ -1,7 +1,7 @@
 #include "album.h"
 
 Album::Album(string Nombre, string ID_Album_, string Fecha,
-             unsigned int Duracion, string Generos, unsigned short Puntuacion, Metrica &x):
+             float Duracion, string Generos, unsigned short Puntuacion, Metrica &x):
     nombre(Nombre), ID_Album(ID_Album_), fecha(Fecha), duracion(Duracion),puntuacion(Puntuacion), generos(Generos){
 
     //portadaRuta="/users/storage/+Artista/+audio/image/"+Nombre;
@@ -14,7 +14,7 @@ Album::~Album(){}
 string Album::getnombre() { return nombre; }
 string Album::getID_Album() { return ID_Album; }
 string Album::getfecha() { return fecha; }
-unsigned int Album::getduracion() { return duracion; }
+float Album::getduracion() { return duracion; }
 string* Album::getgeneros() { return generos; } // devuelve el array directamente
 string Album::getselloDisquero() { return selloDisquero; }
 float Album::getpuntuacion() { return puntuacion; }
@@ -24,7 +24,7 @@ Cancion** Album::getCanciones() { return Canciones; }
 void Album::setnombre(string Nombre) { nombre = Nombre; }
 void Album::setID_Album(string ID) { ID_Album = ID; }
 void Album::setfecha(string Fecha) { fecha = Fecha; }
-void Album::setduracion(unsigned int Duracion) { duracion = Duracion; }
+void Album::setduracion(float Duracion) { duracion = Duracion; }
 void Album::setgeneros(string g1, string g2, string g3, string g4) {
     generos[0] = g1;
     generos[1] = g2;
